compiler errors

+++ src/lib/determ_clock.c
@@ -4,6 +4,7 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <sys/mman.h>
+#include <sys/syscall.h>
 
 #include "determ_clock.h"
 #include "perf_counter.h"
@@ -46,7 +47,7 @@ __attribute__((constructor)) static void determ_clock_init(){
   //initialize the first clock now
   determ_task_clock_init(0);
   //now make a system call to open the task_clock in the kernel
-  __make_clock_sys_call((unsigned long)clock_info->clocks, 0, clock_info->clocks[0].perf_counter->fd);
+  __make_clock_sys_call(clock_info->clocks, 0, clock_info->clocks[0].perf_counter->fd);
   printf("INITIALIZED\n");
 }
